+++ examples/typescript/http-server-pool/express-worker_threads/src/pool.ts
@@ -21,6 +21,6 @@ WorkerResponse<BodyPayload>
     concurrency: 8
   },
   errorHandler: (e: Error) => {
-    console.error(e)
+    console.error('Thread worker error:', e)
   }
 })