+++ examples/typescript/http-server-pool/fastify-hybrid/@types/fastify/index.d.ts
@@ -1,0 +1,17 @@
+import type * as fastify from 'fastify'
+import { type DynamicThreadPool } from 'poolifier'
+import {
+  type ThreadWorkerData,
+  type ThreadWorkerResponse
+} from '../../src/types.ts'
+
+declare module 'fastify' {
+  export interface FastifyInstance extends fastify.FastifyInstance {
+    pool: DynamicThreadPool<ThreadWorkerData, ThreadWorkerResponse>
+    execute: (
+      data?: ThreadWorkerData,
+      name?: string,
+      transferList?: TransferListItem[]
+    ) => Promise<ThreadWorkerResponse>
+  }
+}