+++ examples/typescript/http-server-pool/fastify-hybrid/src/types.ts
@@ -1,0 +1,28 @@
+import type { ThreadPoolOptions } from 'poolifier'
+
+export interface ClusterWorkerData {
+  port: number
+}
+
+export interface ClusterWorkerResponse {
+  status: boolean
+  port?: number
+}
+
+export interface BodyPayload {
+  number?: number
+}
+
+export interface ThreadWorkerData<T = unknown> {
+  body: T
+}
+
+export interface ThreadWorkerResponse<T = unknown> {
+  body: T
+}
+
+export interface FastifyPoolifierOptions extends ThreadPoolOptions {
+  workerFile: string
+  maxWorkers?: number
+  minWorkers?: number
+}