+++ examples/typescript/websocket-server-pool/ws-cluster/src/main.ts
@@ -29,7 +29,7 @@ const pool = new FixedClusterPool<WorkerData, WorkerResponse>(
         })
     },
     errorHandler: (e: Error) => {
-      console.error(e)
+      console.error('Cluster worker error', e)
     }
   }
 )